# Conflicts:
#	roiextractors/extractors/nwbextractors/nwbextractors.py

+++ roiextractors/segmentationextractor.py
@@ -231,7 +231,7 @@ class SegmentationExtractor(ABC, BaseExtractor):
         samp_freq: float
             Sampling frequency of the recordings in Hz.
         """
-        return self._sampling_frequency
+        return self._sampling_frequency.astype(np.float)
 
     def get_num_rois(self):
         """Returns total number of Regions of Interest in the acquired images.
